check for file existance and add vim configs

+++ bin/mac/install-tmux.sh
@@ -3,7 +3,14 @@
 brew update
 brew install tmux
 
-rm -rf ~/.tmux/plugins/tpm
+# clean up previous files
+if [ -e "$HOME/.tmux.conf" ]; then
+  rm ~/.tmux.conf
+fi
+if [ -d "$HOME/.tmux/plugins/tpm" ]; then
+  rm -rf ~/.tmux/plugins/tpm
+fi
+
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
 ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )/../../"
